+++ src/Managers/Minecraft/index.ts
@@ -1,0 +1,3 @@
+export * from './ChatManager';
+export * from './LogManager';
+export * from './StatusManager';
