+++ src/Managers/index.ts
@@ -1,6 +1,4 @@
 export * from './CommandManager';
 export * from './EventManager';
 export * from './LocalizationManager';
-export * from './StatusManager';
-export * from './MinecraftLogManager';
-export * from './MinecraftChatManager';
+export * from './Minecraft';
